Add docker limits example
Revert to docker-compose version 2.0 since that's the official method of imposing limits on docker hosts. If users are familiar with Swarm it should be trivial to swap back to version 3.0. Koji will by default use 100% of all resources on a host which is going to result in a bad time for users that run this on their single VM with RDM, Database, Discord alerts, etc. Capping the speed a little is likely an overall win for _most_ users.

Disabling the limit by default.